// web/src/api/schema.d.ts
export interface components {
    schemas: {
// ...
        SharedFileOut: {
            /** Id */
            id: string;
            /** File Name */
            file_name: string;
            /** Mime */
            mime: string | null;
            /** Size Bytes */
            size_bytes: number | null;
            /** Category */
            category: string;
            /** From Device */
            from_device: string;
            /** Status */
            status: string;
            /** Created At */
            created_at: string;
        };
// ...
    };
    responses: never;
    parameters: never;
    requestBodies: never;
    headers: never;
    pathItems: never;
}